requirements.txt
disnake
requests
pyyaml
jsonc-parser
munch
